agent: Move common unit tests about device
Move common unit tests about device to mod.rs
